abstracted field formatters
