+++ modules/mrc_media/mrc_media.module
@@ -18,6 +18,14 @@ use Drupal\Core\Url;
 /**
  * Implements hook_preprocess_HOOK().
  */
+function mrc_media_preprocess_field(&$variables) {
+//  dpm($variables);
+//  kint($variables);
+}
+
+/**
+ * Implements hook_preprocess_HOOK().
+ */
 function mrc_media_preprocess_dropzonejs(&$variables) {
   // Adds additional things to the template for the dropzone js widget.
   $variables['allowed_files'] = str_replace(' ', ', ', $variables['element']['#extensions']);
@@ -107,15 +115,6 @@ function mrc_media_entity_embed_alter(array &$build, EntityInterface $entity, ar
  * Implements hook_preprocess_HOOK().
  */
 function mrc_media_preprocess_media(&$variables) {
-  if ($variables['media']->bundle() == 'image') {
-    if (isset($variables['elements']['#mrc_media_image_style'])) {
-      MediaImageFormatter::alterMediaRender($variables);
-    }
-    elseif (isset($variables['elements']['#mrc_media_responsive_image_style'])) {
-      MediaResponsiveImageFormatter::alterMediaRender($variables);
-    }
-  }
-
   /** @var \Drupal\media\Entity\MediaType $media_type */
   $media_type = \Drupal::entityTypeManager()->getStorage('media_type')
     ->load($variables['media']->bundle());
