+++ modules/mrc_media/src/Plugin/Field/FieldFormatter/MediaImageFormatter.php
@@ -2,13 +2,7 @@
 
 namespace Drupal\mrc_media\Plugin\Field\FieldFormatter;
 
-use Drupal\Core\Field\BaseFieldDefinition;
-use Drupal\Core\Field\FieldDefinitionInterface;
-use Drupal\Core\Field\FieldItemListInterface;
-use Drupal\Core\Field\Plugin\Field\FieldFormatter\EntityReferenceEntityFormatter;
-use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\Render\Element;
-use Drupal\field\Entity\FieldConfig;
 
 /**
  * Plugin implementation of the 'yearonly_academic' formatter.
@@ -22,106 +16,32 @@ use Drupal\field\Entity\FieldConfig;
  *   }
  * )
  */
-class MediaImageFormatter extends EntityReferenceEntityFormatter {
+class MediaImageFormatter extends MediaFormatter {
 
   /**
    * {@inheritdoc}
    */
-  public static function defaultSettings() {
-    return [
-        'image_style' => NULL,
-        'link' => 0,
-      ] + parent::defaultSettings();
+  protected function getStyleOptions() {
+    return image_style_options(FALSE);
   }
 
   /**
    * {@inheritdoc}
    */
-  public function settingsForm(array $form, FormStateInterface $form_state) {
-    $elements = parent::settingsForm($form, $form_state);
-
-    $elements['image_style'] = [
-      '#type' => 'select',
-      '#options' => image_style_options(FALSE),
-      '#title' => t('Image Style'),
-      '#default_value' => $this->getSetting('image_style') ?: '',
-      '#empty_option' => $this->t('Use Entity Display'),
-    ];
-    $elements['link'] = [
-      '#type' => 'checkbox',
-      '#title' => $this->t('Link Media to Parent'),
-      '#default_value' => $this->getSetting('link'),
-    ];
-    return $elements;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function isApplicable(FieldDefinitionInterface $field_definition) {
-    $is_applicable = parent::isApplicable($field_definition);
-    $target_type = $field_definition->getFieldStorageDefinition()
-      ->getSetting('target_type');
-    return $is_applicable && $target_type == 'media';
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function settingsSummary() {
-    $image_styles = image_style_options(FALSE);
-    $summary = parent::settingsSummary();
-
-    unset($image_styles['']);
-    $image_style_setting = $this->getSetting('image_style');
-    if (isset($image_styles[$image_style_setting])) {
-      $summary[] = t('Image style: @style', ['@style' => $image_styles[$image_style_setting]]);
-    }
-    else {
-      $summary[] = t('Use Entity Display');
-    }
-
-    return $summary;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function viewElements(FieldItemListInterface $items, $langcode) {
-    $elements = parent::viewElements($items, $langcode);
-    $image_styles = image_style_options(FALSE);
-    $style = $this->getSetting('image_style');
-
-    if (empty($style) || !isset($image_styles[$style])) {
-      return $elements;
-    }
-    /** @var \Drupal\Core\Entity\EntityInterface $parent */
-    $parent = $items->getParent()->getValue();
-
-    foreach ($elements as &$element) {
-      $element['#mrc_media_image_style'] = $style;
-
-      if ($this->getSetting('link')) {
-        $element['#mrc_media_url'] = $parent->toUrl();
-      }
-    }
-    return $elements;
-  }
-
-  /**
-   * @param $element
-   */
-  public static function alterMediaRender(&$element){
-    $element['content']['field_media_image']['#formatter'] = 'image';
-    foreach (Element::children($element['content']['field_media_image']) as $delta) {
-      $item = &$element['content']['field_media_image'][$delta];
+  public static function preRender($element) {
+    $element['field_media_image']['#formatter'] = 'image';
+    foreach (Element::children($element['field_media_image']) as $delta) {
+      $item = &$element['field_media_image'][$delta];
       $item['#theme'] = 'image_formatter';
-      $item['#image_style'] = $element['elements']['#mrc_media_image_style'];
+      $item['#image_style'] = $element['#mrc_media_image_style'];
 
-      if (isset($element['elements']['#mrc_media_url'])) {
-        $item['#url'] = $element['elements']['#mrc_media_url'];
+      if (isset($element['#mrc_media_url'])) {
+        $item['#url'] = $element['#mrc_media_url'];
+        $item['#attributes']['title'] = $element['#mrc_media_url_title'];
       }
+
     }
+    return $element;
   }
 
 }
